Refactor: Rename and relocate grid components, adjust page navigation
- Renamed `Grid` to `SimpleGrid` and `GridSubcomposeLayout` to `SimpleGridSubcomposeLayout` for better clarity.
- Relocated `GridItemParentData` and `Grid` to a new package structure.
- Renamed PlaceholderScreen to GridScreen and moved it to the `grid` package
- Removed unnecessary `GridSubcomposeLayout` in `PagerScreen` and `GridScreen`.
- Added `canScroll` to prevent infinite page creation when a user has reached the last page.
- Refactored page navigation to reset to the first or last page when a new page is created.
- Updated import paths to reflect these changes.
