+++ feature/home/src/main/kotlin/com/eblan/launcher/feature/home/model/HomeType.kt
@@ -1,5 +1,5 @@
 package com.eblan.launcher.feature.home.model
 
 enum class HomeType {
-    Pager, Application, Widget, Placeholder
+    Pager, Application, Widget, Grid
 }
