+++ feature/home/src/main/kotlin/com/eblan/launcher/feature/home/screen/grid/Grid.kt
@@ -1,4 +1,4 @@
-package com.eblan.launcher.feature.home.screen.pager.component.grid
+package com.eblan.launcher.feature.home.screen.grid
 
 import androidx.compose.animation.core.animateIntAsState
 import androidx.compose.material3.Surface
