+++ feature/home/src/main/kotlin/com/eblan/launcher/feature/home/screen/grid/GridItemParentData.kt
@@ -1,4 +1,4 @@
-package com.eblan.launcher.feature.home.screen.pager.component.grid
+package com.eblan.launcher.feature.home.screen.grid
 
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.layout.ParentDataModifier
